tests/wpt/web-platform-tests/css/css-flexbox/intrinsic-width-overflow-auto.tentative.html

<!DOCTYPE html>
<link rel="author" title="Google" href="https://www.google.com/" />
<link rel="help" href="https://drafts.csswg.org/css-flexbox-1/#intrinsic-sizes" />
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/1865" />

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>

<body onload="checkLayout('.flexbox')">
<div class="flexbox" style="display: flex; width: min-content;" data-expected-width="0">
  <div style="overflow: auto;">
    <div style="width: 100px; height: 100px;"></div>
  </div>
</div>

<div class="flexbox" style="display: flex; width: min-content;" data-expected-width="10">
  <div style="overflow: auto; border: 5px solid;">
    <div style="width: 100px; height: 100px;"></div>
  </div>
</div>
